requirements.txt
streamlit
ultralytics
opencv-python
numpy
pillow
matplotlib
plotly
pyyaml
pandas 